Document parsing mostly works

+++ DocumentProcessor/Common/Models/DocumentData.cs
@@ -9,10 +9,10 @@ namespace Common.Models
         public int FileSize { get; set; }
         public string VendorName { get; set; }
         public string InvoiceDate { get; set; }
-        public int TotalAmount { get; set; }
-        public int TotalAmountDue { get; set; }
+        public decimal TotalAmount { get; set; }
+        public decimal TotalAmountDue { get; set; }
         public string Currency { get; set; }
-        public int TaxAmount { get; set; }
+        public decimal TaxAmount { get; set; }
         public string ProcessingStatus { get; set; }
         public string Id { get; set; }
     }
